requirements.txt
discord.py
aiocron
flask
